done hw

+++ controllers/contacts/removeById.js
@@ -3,7 +3,8 @@ const { NotFound } = require("http-errors");
 
 const removeById = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findByIdAndRemove(id);
+  const { _id } = req.user;
+  const result = await Contact.findByIdAndRemove({ _id: id, owner: _id });
   if (!result) {
     throw new NotFound(`Contact wit id=${id} not found`);
   }
